[resolver] Move build resolution methods to BuildResolver
